requirements.txt
streamlit
google-generativeai
PyPDF2
pytesseract
Pillow
